// migration-newip-to-ts/src/components/utils/enums.ts
export enum EnumSource {
  item = 'source__item',
  sourceId = 'data-source-id',
  source = 'data-source',
}

export enum EnumNews {
  item = '.news__item',
  photo = '.news__meta-photo',
  author = '.news__meta-author',
  date = '.news__meta-date',
  title = '.news__description-title',
  source = '.news__description-source',
  content = '.news__description-content',
  more = '.news__read-more a',
  news = '.news',
}